Moved WebApplicationBuilderConfiguration to a separate file, changed PasswordService scope to Singleton
